Moved table class to display
