+++ openatlas/views/admin.py
@@ -18,6 +18,7 @@ from wtforms.validators import InputRequired
 from openatlas import app
 from openatlas.database.connect import Transaction
 from openatlas.display.tab import Tab
+from openatlas.display.table import Table
 from openatlas.forms.setting import (
     ApiForm, ContentForm, FilesForm, GeneralForm, LogForm, MailForm, MapForm,
     ModulesForm, SimilarForm, TestMailForm)
@@ -32,7 +33,6 @@ from openatlas.models.type import Type
 from openatlas.models.user import User
 from openatlas.util.image_processing import (
     create_resized_images, delete_orphaned_resized_images)
-from openatlas.util.table import Table
 from openatlas.util.util import (
     button, convert_size, display_form, display_info, format_date,
     get_file_path, is_authorized, link, manual, required_group, sanitize,
