+++ openatlas/views/export.py
@@ -10,8 +10,8 @@ from werkzeug.wrappers import Response
 from wtforms import SubmitField
 
 from openatlas import app
+from openatlas.display.table import Table
 from openatlas.models.export import sql_export
-from openatlas.util.table import Table
 from openatlas.util.util import (
     convert_size, is_authorized, link, required_group, uc_first)
