+++ openatlas/views/hierarchy.py
@@ -7,10 +7,10 @@ from werkzeug.wrappers import Response
 
 from openatlas import app
 from openatlas.database.connect import Transaction
+from openatlas.display.table import Table
 from openatlas.forms.form import get_manager
 from openatlas.models.entity import Entity
 from openatlas.models.type import Type
-from openatlas.util.table import Table
 from openatlas.util.util import (
     display_form, get_entities_linked_to_type_recursive, link, required_group,
     sanitize, uc_first)
