+++ openatlas/views/imports.py
@@ -16,9 +16,9 @@ from wtforms.validators import InputRequired
 from openatlas import app
 from openatlas.database.connect import Transaction
 from openatlas.display.tab import Tab
+from openatlas.display.table import Table
 from openatlas.models.entity import Entity
 from openatlas.models.imports import Import, is_float
-from openatlas.util.table import Table
 from openatlas.util.util import (
     button, datetime64_to_timestamp, display_form, format_date,
     get_backup_file_data,
